rename folder to graphql
