+++ packages/github-api/src/index.ts
@@ -1,6 +1,6 @@
 import { createFoundationSimulationServer } from "@simulacrum/foundation-simulator";
 import { extendStore } from "./store/index";
-import { extendRouter } from "./service/extend-api";
+import { extendRouter } from "./graphql/extend-api";
 
 export const simulation = createFoundationSimulationServer({
   port: 3300, // default port
